+++ AddInManager/Command/AddinManagerBase.cs
@@ -14,7 +14,7 @@ public sealed class AddinManagerBase
 {
     public Result ExecuteCommand(ExternalCommandData data, ref string message, ElementSet elements, bool faceless)
     {
-        var vm = new AddInManagerViewModel(data);
+        var vm = new AddInManagerViewModel(data,ref message,elements);
         if (_mActiveCmd != null && faceless)
         {
             return RunActiveCommand(vm, data, ref message, elements);
@@ -23,11 +23,7 @@ public sealed class AddinManagerBase
         frmAddInManager.WindowStartupLocation = WindowStartupLocation.CenterOwner;
         var process = Process.GetCurrentProcess();
         new WindowInteropHelper(frmAddInManager).Owner = process.MainWindowHandle;
-        var showDialog = frmAddInManager.ShowDialog();
-        if (showDialog == false && ActiveCmd != null&& vm.IsRun)
-        {
-            return RunActiveCommand(vm, data, ref message, elements);
-        }
+        frmAddInManager.Show();
         return Result.Failed;
     }
 
@@ -38,7 +34,7 @@ public sealed class AddinManagerBase
     }
 
 
-    private Result RunActiveCommand(AddInManagerViewModel vm, ExternalCommandData data, ref string message, ElementSet elements)
+    public Result RunActiveCommand(AddInManagerViewModel vm, ExternalCommandData data, ref string message, ElementSet elements)
     {
         var filePath = _mActiveCmd.FilePath;
         Result result;