+++ AddInManager/ViewModel/AddInManagerViewModel.cs
@@ -4,6 +4,7 @@ using System.IO;
 using System.Windows;
 using System.Windows.Input;
 using System.Xml;
+using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
 using RevitAddinManager.Command;
 using RevitAddinManager.Model;
@@ -14,8 +15,8 @@ namespace RevitAddinManager.ViewModel;
 
 public class AddInManagerViewModel : ViewModelBase
 {
-    public bool IsRun { get; set; }
     public ExternalCommandData ExternalCommandData { get; set; }
+    private ElementSet Elements { get; set; }
     public View.FrmAddInManager FrmAddInManager { get; set; }
     public AssemLoader AssemLoader { get; set; }
 
@@ -156,7 +157,7 @@ public class AddInManagerViewModel : ViewModelBase
         set => OnPropertyChanged(ref _vendorDescription, value);
     }
 
-    private bool _issTabCmdSelected;
+    private bool _issTabCmdSelected = true;
     public bool IsTabCmdSelected
     {
         get => _issTabCmdSelected;
@@ -193,15 +194,19 @@ public class AddInManagerViewModel : ViewModelBase
         Process.Start("https://github.com/chuongmep/RevitAddInManager/wiki");
     }
 
-    public AddInManagerViewModel(ExternalCommandData data)
+    public AddInManagerViewModel(ExternalCommandData data,ref string message,ElementSet elements)
     {
         AssemLoader = new AssemLoader();
         MAddinManagerBase = AddinManagerBase.Instance;
         CommandItems = FreshTreeItems(false, MAddinManagerBase.AddinManager.Commands);
         ApplicationItems = FreshTreeItems(false, MAddinManagerBase.AddinManager.Applications);
         ExternalCommandData = data;
+        Messages = message;
+        Elements = elements;
         FreshItemStartupClick(false);
     }
+
+    public string Messages { get; set; }
 
     public ObservableCollection<AddinModel> FreshTreeItems(bool isSearchText, Addins addins)
     {
@@ -254,6 +259,7 @@ public class AddInManagerViewModel : ViewModelBase
     {
         try
         {
+            if(FrmAddInManager==null) return;
             if (SelectedCommandItem?.IsParentTree == false)
             {
                 MAddinManagerBase.ActiveCmd = SelectedCommandItem.Addin;
@@ -261,8 +267,15 @@ public class AddInManagerViewModel : ViewModelBase
                 CheckCountSelected(CommandItems, out var result);
                 if (result > 0)
                 {
-                    IsRun = true;
-                    FrmAddInManager.Close();
+                    
+                    if (MAddinManagerBase.ActiveCmd!=null)
+                    {
+                        FrmAddInManager.Close();
+                        FrmAddInManager = null;
+                        string messages = Messages;
+                        MAddinManagerBase.RunActiveCommand(this, ExternalCommandData,ref messages, Elements);
+                    }
+                   
                 }
             }